\name{akima}
\alias{akima}
\title{
Waveform Distortion Data for Bivariate Interpolation
}
\arguments{
\item{x,y,z}{
represents a smooth surface of \code{z} values at selected points
irregularly distributed in the \code{x-y} plane.
}}
\description{
Hiroshi Akima, "A Method of Bivariate Interpolation and
Smooth Surface Fitting for Irregularly Distributed Data
Points", 
ACM Transactions on Mathematical Software,
Vol. 4, No. 2, June 1978, pp. 148-159.
Copyright 1978, Association for Computing Machinery, Inc.,
reprinted by permission.


The data was taken from a study of waveform distortion in
electronic circuits, described in:
Hiroshi Akima, "A Method of Bivariate Interpolation and
Smooth Surface Fitting Based on Local Procedures", 
CACM,
Vol. 17, No. 1, January 1974, pp. 18-20.
}
\keyword{datasets}
% Converted by Sd2Rd version 0.2-a3.
